frontend/core/blocks/watermarking/requirements.txt
transformers==4.38.1
numpy==1.26.4
scikit-learn==1.4.1.post1
pandas==2.0.3
transformers==4.38.1
Pillow==10.4.0
tqdm==4.66.4
matplotlib==3.9.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5